Add check whos article is, to show correspond btns

+++ app/helpers/articles_helper.rb
@@ -2,9 +2,10 @@ module ArticlesHelper
 
 
   def show_edit_del_btns article
-    result = link_to ' | Edit', edit_article_path(article) 
-    result += link_to ' | Delete', article, data: { :confirm => 'Are you sure?'}, :method => :delete
-    result.html_safe
+    if user_signed_in? && article.title == current_user.username
+      result = link_to ' | Edit', edit_article_path(article) 
+      result += link_to ' | Delete', article, data: { :confirm => 'Are you sure?'}, :method => :delete
+      return result.html_safe
+    end
   end
-
 end
